Rename codeActionMenu to codeActionWidget (#161540)
