+++ src/vs/editor/contrib/codeAction/browser/codeActionWidget.css
